Exercise 4.21 completion (bloglist)

+++ controllers/blogs.js
@@ -29,6 +29,14 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   const id = request.params.id
+  const blog = await Blog.findById(id)
+  const decodedToken = jwt.decode(request.body.token, process.env.SECRET)
+  if(!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  if(decodedToken.id.toString() !== blog.user._id.toString()) {
+    return response.status(401).json({ error: 'unauthorized user' })
+  }
   const result = await Blog.findByIdAndDelete(id)
   response.status(204).send(result)
 })
